requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
coards==1.0.5
docopt==0.6.2
Extract==1.0
idna==3.2
inflect==5.3.0
Jinja2==3.0.1
json-extract==1.1.2
MarkupSafe==2.0.1
nose==1.3.7
numpy==1.21.1
Pydap==3.2.2
python-dotenv==0.19.0
PyYAML==5.4.1
Random-Word==1.0.7
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
WebOb==1.8.7
